feat: delete de comment

+++ 02-Acme_Antenna/Workspace/Acme-Antenna/src/main/java/controllers/administrator/CommentAdministratorController.java
@@ -31,20 +31,29 @@ public class CommentAdministratorController extends AbstractController{
 		Comment comment;
 
 		comment = this.commentService.findOne(commentId);
+		System.out.println("Comment en edit: "+comment);
 		result = this.createEditModelAndView(comment);
 
 		return result;
 	}
 	@RequestMapping(value = "/edit", method = RequestMethod.POST, params = "delete")
 	public ModelAndView delete(@Valid final Comment comment, final BindingResult binding) {
+		System.out.println("Comment en delete: " + comment);
 		ModelAndView res;
-		try {
-			this.commentService.delete(comment);
-			res = new ModelAndView("redirect:../../");
-		} catch (final Throwable oops) {
-			System.out.println(oops.getMessage());
-			res = this.createEditModelAndView(comment, "comment.commit.error");
+		if (binding.hasErrors()) {
+			System.out.println(binding.getAllErrors());
+			res = this.createEditModelAndView(comment, "comment.params.error");
+		} else {
+			try {
+				this.commentService.delete(comment);
+				res = new ModelAndView("redirect:../../");
+			} catch (final Throwable oops) {
+				System.out.println(oops.getMessage());
+				System.out.println(oops.getStackTrace());
+				res = this.createEditModelAndView(comment, "comment.commit.error");
+			}
 		}
+		
 		return res;
 	}
 	
@@ -59,7 +68,7 @@ public class CommentAdministratorController extends AbstractController{
 	protected ModelAndView createEditModelAndView(final Comment comment, final String message) {
 		ModelAndView result;
 
-		result = new ModelAndView("comment/edit");
+		result = new ModelAndView("comment/administrator/edit");
 		result.addObject("comment", comment);
 		result.addObject("message", message);
 		result.addObject("requestUri", "comment/administrator/edit.do");
